+++ CHANGELOG.md
@@ -12,6 +12,9 @@ glance - what to expact from upgrading to a new version.
 
 ### Fixed
 
+- Removes seconds part of output of `timespan_in_words` in single-unit mode due
+  to floating point inaccuracies.
+
 ### Security
 
 ### Deprecated